requirements.txt
fastapi
uvicorn[standard]
crewai
langchain_community
langchain_ollama
duckduckgo-search
python-dotenv
sqlalchemy
aiosqlite
